Доработка экспорта в промед стационарных случаев

+++ ecom/ecom-riams/ejb/src/main/java/ru/ecom/mis/ejb/domain/lpu/MisLpu.java
@@ -283,7 +283,7 @@ public class MisLpu extends BaseEntity {
     /**
      * Ид отделения в промеде
      */
-    private Integer promedLpuSectionId;
+    private Long promedLpuSectionId;
 
 
     /**
